move to method to the interface
